refactor: Turn 리팩토링
